requirements.txt
Flask==1.0.2
uWSGI==2.0.17.1
Pillow==5.4.0
